requirements.txt
# ============================================================
# HybridRAG Bench — Pinned Dependencies
# Python 3.11+
# ============================================================

# ── Core ML / Embeddings ──────────────────────────────────────
sentence-transformers==3.4.1
torch==2.1.2
transformers==4.41.0

# ── Vector Store ──────────────────────────────────────────────
qdrant-client==1.17.1

# ── Lexical Retrieval ─────────────────────────────────────────
rank-bm25==0.2.2

# ── NLP Utilities ─────────────────────────────────────────────
nltk==3.8.1

# ── Evaluation Metrics ────────────────────────────────────────
rouge-score==0.1.2
numpy==1.26.4
pandas==2.0.3

# ── LLM Providers ─────────────────────────────────────────────
google-generativeai==0.8.6
openai==1.51.0

# ── API Service ───────────────────────────────────────────────
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1

# ── Dashboard ─────────────────────────────────────────────────
streamlit==1.39.0
plotly==5.24.0

# ── Configuration ─────────────────────────────────────────────
pyyaml==6.0.2

# ── Logging ───────────────────────────────────────────────────
colorlog==6.8.2

# ── Utilities ─────────────────────────────────────────────────
tqdm==4.66.5
colorama==0.4.6

# ── Testing ───────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2      # for FastAPI TestClient
